Backend/requirements.txt
alembic==1.15.2
authlib==1.5.1
cryptography==36.0.0
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.5
openai==1.61.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==20.1.0